* 6.4:
  [BeanstalkMessenger] Round delay to an integer to avoid deprecation warning
  [PropertyInfo] Fix interface handling in `PhpStanTypeHelper`
  [HttpClient] Test POST to GET redirects
  [HttpKernel] Denormalize request data using the csv format when using "#[MapQueryString]" or "#[MapRequestPayload]" (except for content data)
  fix: preserve and nowrap in profiler code highlighting

+++ Transport/Connection.php
@@ -124,7 +124,7 @@ class Connection
             $job = $this->client->useTube($this->tube)->put(
                 $message,
                 PheanstalkInterface::DEFAULT_PRIORITY,
-                $delay / 1000,
+                (int) ($delay / 1000),
                 $this->ttr
             );
         } catch (Exception $exception) {
